requirements.txt
flask
facenet-pytorch
torch
torchvision
scikit-learn
pillow
werkzeug